Move handshake types to separate file
